Вынес Singleton в трейт
